simplify: Minor refactoring of connected components
Add a few comments to make the code easier to read, and simplify the
component update.

Because the last pass renumbers the components which changes the meaning
of the values from parent links to component ids, we *need* an
intermediate pass to finalize component values; make that clear in
comments.