simplify: Implement component pruning
When pruning is enabled, we perform a second pass filtering of the index
buffer using the current reached error limit; any triangles that are
part of components with a smaller error get removed.

Note that this is done for all components regardless of whether or not
some of them are unnecessary to remove to reach the resulting count (but
only done if we haven't reached the count yet). This means we do not
need to sort the components by error but may result in bursts of triangle
counts depending on how the edge collapses end up working out.