demo: Add a basic test for pruning
Here we have one degenerate triangle that has all three vertices locked
(and is not topologically degenerate); this triangle can not be
collapsed normally, however pruning collapses it as it has zero error,
just as the other collapses here.

We also check that the resulting small mesh can be processed with
pruning enabled without triggering any assertions.