simplify: Iteratively prune components until error limit
In addition to pruning after each pass, where we are limiting the error
by the current worst-case error reached, we now also prune components at
the end of simplification with a series of passes of increasing error.

This helps on meshes where available edge collapses are very low error
so the total error must increase to prune components; without this
meshes would be stuck at a higher triangle count. This also improves the
behavior with varying target_error: if some edge collapses have error
1.1e-2, and some components have an error 0.9e-2, then without this we
might see components pruned with target error 1.2e-2 but not with 1e-2
because earlier edges might have lower error.