simplify: Remove a seemingly redundant call to pruneComponents
This was originally added so that we have a pruneComponents call at the
end of any simplification process. However, since pruneComponents will
only do work if the error has increased since last time, and early-outs
from the edge collapse loop are only possible if the error has not
updated, even before cleanup passes this should have been redundant with
a rare exception of pruning zero-error components in a fully locked mesh.

With the cleanup passes this is entirely redundant now, and is never hit
on any real test meshes.