vertexcodec: Add a note about quantization and locality of reference
For optimum compression, the vertex buffer should carry quantized values
and have consecutive vertices be reasonably local which can be achieved
with regular vertex cache/fetch optimization for most meshes.

An alternative to quantization is filtering but since it is unlikely to
be as broadly used it's probably enough to just call out quantization
here.