demo: Implement manual boundary locking for clusters
While using automatic border locking usually works very well, it
prevents simplifying open geometric borders that the source mesh might
contain. When a mesh has many edges like this, this negatively affects
the DAG quality by locking edges contained within one cluster group.

Instead, we can simply mark vertices that are shared by more than one
group after a partition and feed that data to simplifier.

The effect of this is situational; it's beneficial overall but because
the rest of the pipeline has odd issues with partitioning, even when
using METIS for everything, sometimes this ends up constraining the
build process a little more for unclear reasons. Theoretically this
should be strictly better though, and on some meshes like kitten.obj it
does measurably help, however it needs fixes for attribute seams to
keep it enabled.