demo: Use position remap for locking and partitioning
Previously when we built adjacency between groups we would not create
edges between clusters when they shared the position of the vertex with
a different attribute. This meant that when the cluster boundary
contained a crease, it would not be reflected correctly in the
adjacency.

Similarly, but more importantly, we need this information when locking
the cluster boundary for the same reason; before, lockBoundary could
result in gaps because the locking information on the boundary was
inconsistent. This allows us to enable kUseLocks by default.