demo: Use a flexible (and larger) cluster group size
Using 4 clusters is fairly limiting wrt partition quality, as it demands
very careful grouping that METIS is not willing to provide. Instead use
a larger group size (8 atm), and adjust the simplification condition to
be adaptive to the group size so that partitioning is free to produce
groups of varying sizes.