tools: Implement round-trip fuzzing for vertex codec
In addition to validating all decoders for memory safety we can also
validate that the vertex encoder in particular round-trips safely. Index
encoder may rotate triangles so it's not a perfect binary match, and
index sequence encoder is rarely used and very simple, but vertex
encoder is more complicated and may become more so with v1.