gltfpack: Implement computation of mesh geometry hash
To make it possible to deduplicate meshes, but not have to do a pairwise
compare of the entire mesh contents, we can precompute a 128-bit hash
(using MurMur3, chosen for the simplicity of implementation and good
collision properties); this change does that without the deduplication
code for now.

Calculating the hashes is reasonably fast: it takes ~2% of processing
time for geometry heavy scenes give or take.