gltfpack: Record the fact that mesh geometry might be duplicate
Instead of computing a hash after mesh processing for every mesh, we now
only do this when the mesh has been tagged as possibly-duplicate. This
reduces the cost of hash update to cases when we are likely to benefit
from them, and makes it easier to disable deduplication entirely for
testing.

While it is possible that two meshes did not have duplicate geometry
before processing and are processed to have it (eg two identical meshes
with a redundant vertex stream on one but not the other that gets
removed), this is fairly unlikely in practice.