gltfpack: Avoid detaching meshes with duplicate geometry
When a mesh has duplicate geometry, detaching it from its node will
break the data sharing because of the world-space transform. When -mm is
specified, this is fair game: it's a request to trade off draw call
count for data size, but by default it would be more consistent to leave
these as is.

This can increase the number of used nodes and mesh objects, but not
beyond what we already use with `-kn`, and the extra cost generally is
paid off by the savings in deduplicated data.