gltfpack: Add temporary command line option to disable mesh deduplication
In theory, deduplication should not result in any issues as any changes
only concern the internal file structure, and it should be beneficial in
most cases (and when it's penalizing draw call counts, existing options
provide enough control), but just in case we now can disable this
feature entirely using -mdd.

The goal is to remove this option in the future, it's just here in case
anyone reports issues in the wild.

Not running dedupMeshes means no mesh has geometry_duplicate flag set,
so no dedup related code activates.