simplify: Use error returned by vertex quadrics to guide pruning
When attributes are incorporated into the metric, depending on the
attribute weights we may see divergence between the error as returned by
vertex quadrics alone and the result. Especially when the simplification
does not have a set limit, this may lead to unexpectedly aggressive
component pruning.

While ideally pruning should only be used with a reasonable error limit
that negates these concerns, for now it is probably better to accomodate
this use case by tracking the positional error separately.

To reduce the cost of tracking this error we recompute it as collapses
are performed; this may seem counter-intuitive since we've already
computed it during ranking, but conserving memory is more important
here, and since we only need to do it when using attributes and only for
collapses we end up taking, the performance cost ends up being
negligible.