gltfpack: Avoid searching for duplicate geometry when there's none
While on most scenes our strategy performs reasonably well, and we
already have some potential N^2 processing elsewhere, there are cases
where the existing N^2 algorithms in the pipeline are actually
semi-linear, because mesh merging lazily and effectively aggregates most
meshes; since deduplication is the first pass, it may actually be
quadratic in the face of a scene with a lot of meshes.

A simple fix is to track the number of occurences of each hash and only
search for duplicates for meshes where this value is >1. We could of
course implement more involved tracking but for now this works just
fine.