gltfpack: Filter duplicate nodes after merging node lists
This almost never comes up in practice, but occasionally scenes might
have the same geometry attached to the same nodes. This results in
counterintuitive results because we create node lists with duplicate
nodes for this, which never happened before, and end up producing more
quantization sub-nodes than necessary.

For now we can just filter these attachments out to avoid redundancy.