demo: Add average strip length to demo output
This is a similar efficiency measure to strip index count and can be
used to compare the results to other published algorithms. For
simplicity, we only measure the strip length when restart indices are
used; the results should not depend on whether that's the case (compared
to strip index percentage which does).