demo: Adjust "not full" calculation for uniform meshlets
This might be a little misleading in the future, but uniform meshlets
can almost never reach full vertex utilization as the triangle count
should be reached first on regular meshes. So for now it would be more
useful to count meshlets that don't exhaust the triangle limit even
though this makes the metric differ between two types.