js: Fix vertex_lock support for simplifyWithAttributes
This code never worked if an actual array was passed in as it required
an element per float instead of an element per vertex.

Since we need to change this anyway, change the interface to accept
Uint8Array instead of a boolean array. This is more consistent with
other parameters and allows an upgrade path if the C interface changes
from "lock flags" to just "flags".